@@ -2,7 +2,7 @@
     "name": "pitoncms/session",
     "description": "Manage HTTP session state without native PHP sessions",
     "type": "library",
-    "license": "MIT",
+    "license": "AGPL-3.0-or-later",
     "homepage": "https://github.com/pitoncms/session",
     "keywords": ["piton","pitoncms","session","sessions","session state"],
     "authors": [
@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": ">=7.3"
+        "php": ">=8.3"
         },
     "autoload": {
         "psr-4": {
